refactor: defines the base structure of dfget(GoLang)
