+++ dfget/cli/cli.go
@@ -14,18 +14,7 @@
  * limitations under the License.
  */
 
-package main
-
-import (
-	"fmt"
-
-	"github.com/alibaba/Dragonfly/version"
-)
-
-func main() {
-	// TODO: refactor 'dfget' with GoLang
-	// the current dfget is written by python in 'src/getter'
-	// This main function is just the entry of dfget, the other operations
-	// must be controlled in package CLI.
-	fmt.Println("dfget version: " + version.DFGetVersion)
-}
+// Package cli controls the flow of all procedures. It parses arguments,
+// invokes registrar to register itself on supernode, assigns tasks to
+// downloader and other modules.
+package cli
