+++ dfget/config/config.go
@@ -14,18 +14,5 @@
  * limitations under the License.
  */
 
-package main
-
-import (
-	"fmt"
-
-	"github.com/alibaba/Dragonfly/version"
-)
-
-func main() {
-	// TODO: refactor 'dfget' with GoLang
-	// the current dfget is written by python in 'src/getter'
-	// This main function is just the entry of dfget, the other operations
-	// must be controlled in package CLI.
-	fmt.Println("dfget version: " + version.DFGetVersion)
-}
+// Package config holds all properties of dfget.
+package config
