+++ dfget/uploader/uploader.go
@@ -14,18 +14,7 @@
  * limitations under the License.
  */
 
-package main
-
-import (
-	"fmt"
-
-	"github.com/alibaba/Dragonfly/version"
-)
-
-func main() {
-	// TODO: refactor 'dfget' with GoLang
-	// the current dfget is written by python in 'src/getter'
-	// This main function is just the entry of dfget, the other operations
-	// must be controlled in package CLI.
-	fmt.Println("dfget version: " + version.DFGetVersion)
-}
+// Package uploader implements an uploader server. It is the important role
+// - peer - in P2P pattern that will wait for other P2PDownloader to download
+// its downloaded files.
+package uploader
